Eliminate WebKit2 compilation warnings. https://bugs.webkit.org/show_bug.cgi?id=65025
Warnings found with gcc on linux.

Patch by Wajahat Siddiqui <mdwajahatali.siddiqui@motorola.com> on 2011-09-14
Reviewed by Sam Weinig.

* Platform/CoreIPC/ArgumentEncoder.cpp:
(CoreIPC::ArgumentEncoder::~ArgumentEncoder):
* WebProcess/WebCoreSupport/WebFrameLoaderClient.cpp:
(WebKit::WebFrameLoaderClient::createPlugin):

Canonical link: https://commits.webkit.org/84036@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@95154 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit2/Platform/CoreIPC/ArgumentEncoder.cpp
@@ -54,7 +54,7 @@ ArgumentEncoder::~ArgumentEncoder()
 #if !USE(UNIX_DOMAIN_SOCKETS)
     // FIXME: We need to dispose of the attachments in cases of failure.
 #else
-    for (int i = 0; i < m_attachments.size(); ++i)
+    for (size_t i = 0; i < m_attachments.size(); ++i)
         m_attachments[i].dispose();
 #endif
 }
